+++ R/aplot.R
@@ -28,9 +28,25 @@ print.aplot <- function(x, ...) {
     grid.draw(x)
 }
 
-as.patchwork <- function(x) {
-    if (!inherits(x, 'aplot')) {
-        stop("only aplot object supported")
+##' as.patchwork
+##' 
+##' @param x object
+##' @param align "x","y","xy","none", align the axis of x/y or not.
+##' @importFrom patchwork plot_layout
+##' @importFrom ggplot2 ggplotGrob
+##' @export
+as.patchwork <- function(x,
+                         align = "xy") {
+    
+    align <- match.arg(align, c("x","y","xy","none"))
+    
+    if (!inherits(x, 'aplot') && !inherits(x, "gglist")) {
+        stop("only aplot or gglist object supported")
+    }
+    if (inherits(x, "gglist")) {
+        y <- c(list(gglist=x), attr(x, 'params'))
+        res <- do.call(plot_list2, y)
+        return(res)
     }
     
     mp <- x$plotlist[[1]]
@@ -38,15 +54,20 @@ as.patchwork <- function(x) {
         return(ggplotGrob(mp))
     }
     
-    for (i in x$layout[, x$main_col]) {
-        if (is.na(i)) next
-        if (i == 1) next
-        x$plotlist[[i]] <- suppressMessages(x$plotlist[[i]] + xlim2(mp))
+    if(align == "x" || align == "xy"){
+        for (i in x$layout[, x$main_col]) {
+            if (is.na(i)) next
+            if (i == 1) next
+            x$plotlist[[i]] <- suppressMessages(x$plotlist[[i]] + xlim2(mp))
+        }
     }
-    for (i in x$layout[x$main_row,]) {
-        if(is.na(i)) next
-        if (i == 1) next
-        x$plotlist[[i]] <- suppressMessages(x$plotlist[[i]] + ylim2(mp))
+    
+    if(align == "y" || align == "xy"){
+        for (i in x$layout[x$main_row,]) {
+            if(is.na(i)) next
+            if (i == 1) next
+            x$plotlist[[i]] <- suppressMessages(x$plotlist[[i]] + ylim2(mp))
+        }
     }
     
     idx <- as.vector(x$layout)
@@ -67,6 +88,8 @@ as.patchwork <- function(x) {
                      heights= x$height,
                      guides = guides)
 }
+
+
 ##' @importFrom ggplot2 ggplotGrob
 ##' @importFrom patchwork patchworkGrob
 aplotGrob <- function(x) {