+++ R/plot-list.R
@@ -14,6 +14,7 @@
 ##' @param tag_levels format to label plots, will be disable if 'labels' is not NULL
 ##' @param tag_size size of tags
 ##' @param design specification of the location of areas in the layout
+##' @param output one of 'gglist' or 'patchwork'
 ##' @return composite plot
 ##' @importFrom patchwork plot_layout
 ##' @importFrom patchwork plot_annotation
@@ -35,7 +36,10 @@ plot_list <- function(..., gglist = NULL,
                       labels = NULL,        
                       tag_levels = NULL,
                       tag_size = 14,
-                      design = NULL) {
+                      design = NULL, 
+                      output = "patchwork") {
+    
+    output <- match.arg(output, c("gglist", "patchwork"))
 
     gglist <- c(list(...), gglist)
     name <- names(gglist)
@@ -72,7 +76,7 @@ plot_list <- function(..., gglist = NULL,
         }
     }
 
-    gglist(gglist = gglist,
+    res <- gglist(gglist = gglist,
           ncol = ncol, 
           nrow = nrow, 
           byrow = byrow,
@@ -83,7 +87,11 @@ plot_list <- function(..., gglist = NULL,
           tag_levels = tag_levels,
           tag_size = tag_size,
           design = design)        
+    if (output == 'gglist') return(res)
+    as.patchwork(res)
 }
+
+
 
 
 plot_list2 <- function(gglist = NULL,
@@ -160,7 +168,6 @@ gglist <- function(gglist, ...) {
 ##' @method print gglist
 ##' @export
 print.gglist <- function(x, ...) {
-    y = c(list(gglist=x), attr(x, 'params'))
-    print(do.call(plot_list2, y))
+    print(as.patchwork(x))
 }